+++ crates/art/src/types/node_index.rs
@@ -5,10 +5,10 @@ use serde::{Deserialize, Serialize};
 pub enum NodeIndex {
     /// Sequence number of a node in a tree. The root is 1, his children are 2 and 3, and so on,
     /// down to leaves.
-    Index(u32),
+    Index(u64),
     /// Level (starting from root as 0) and position on level (starting from left as 0) of a node
     /// in a tree
-    Coordinate(u32, u32),
+    Coordinate(u64, u64),
     /// Path from the root to the node
     Direction(Vec<Direction>),
 }
